backend/requirements.txt
transformers
sentence-transformers
flask
flask_cors
psycopg2-binary
dotenv